Finalized validation of the decimal and finished Phase 4.

+++ TheHeist/Team/TeamMember.cs
@@ -11,11 +11,11 @@ namespace TheHeist.Team
         //properties:
         public string Name { get; set; }
         public int SkillLevel { get; set; } //how can I force it to be a positive number? Is byte an ok way to do that? Doesn't seem to do it!! Even if the max then has to be 255?
-        public int CourageFactor { get; set; } //how can I force a range? AND - how can I store it as a decimal from the Console user entry??
+        public decimal CourageFactor { get; set; } //how can I force a range? AND - how can I store it as a decimal from the Console user entry??
         //[Range(typeof(decimal), "0.0", "2.0")]
 
         //methods:
-        public void DisplayTeamMember(string name, int skillLevel, int courageFactor)
+        public void DisplayTeamMember(string name, int skillLevel, decimal courageFactor)
         {
             Name = name;
             SkillLevel = skillLevel;
